録音データのファイルパスの管理を行うクラスを作成
